Predict Message AUT Qs test
It works!! : Answer:
AIMessage(content='To make a study application at AUT, you would need to visit the AUT website and navigate to the section for prospective students. There, you can find information on how to apply for your desired program. Typically, you would need to create an account on the AUT online application portal, fill out the application form, upload any required documents (such as academic transcripts, CV, and proof of English proficiency), and pay the application fee if applicable.\n\nThe time it takes to receive an offer of place can vary depending on the program you are applying for, the volume of applications being processed, and the time of year. In general, it can take anywhere from a few weeks to a couple of months to receive an offer of place after submitting your application.\n\nFor the postgraduate diploma in health science in child health at AUT, the program typically consists of 120 points. This means that students would need to complete a total of 120 points worth of courses to successfully graduate from the program. The exact number of courses and points required may vary slightly depending on the specific courses offered in any given semester, but 120 points is the standard requirement for this program.')